+++ model/experiment_list.py
@@ -1,15 +1,14 @@
 from __future__ import absolute_import, division, print_function
 
 from builtins import range
-import collections
 import copy
 import json
 import os
 import pkg_resources
 from typing import Any, Dict, Optional, Tuple
+import warnings
 
 from dxtbx.datablock import (
-    AutoEncoder,
     BeamComparison,
     DataBlockFactory,
     DataBlockTemplateImporter,
@@ -419,135 +418,16 @@ class ExperimentListDumper(object):
 
     def __init__(self, experiment_list):
         """ Initialise """
-        self._experiment_list = experiment_list
-
-    def as_json(self, filename=None, compact=False, split=False):
-        """ Dump experiment list as json """
-        # Get the dictionary and get the JSON string
-        dictionary = self._experiment_list.to_dict()
-
-        # Split into separate files
-        if filename is not None and split:
-
-            # Get lists of models by filename
-            basepath = os.path.splitext(filename)[0]
-            ilist = [
-                ("%s_imageset_%d.json" % (basepath, i), d)
-                for i, d in enumerate(dictionary["imageset"])
-            ]
-            blist = [
-                ("%s_beam_%d.json" % (basepath, i), d)
-                for i, d in enumerate(dictionary["beam"])
-            ]
-            dlist = [
-                ("%s_detector_%d.json" % (basepath, i), d)
-                for i, d in enumerate(dictionary["detector"])
-            ]
-            glist = [
-                ("%s_goniometer_%d.json" % (basepath, i), d)
-                for i, d in enumerate(dictionary["goniometer"])
-            ]
-            slist = [
-                ("%s_scan_%d.json" % (basepath, i), d)
-                for i, d in enumerate(dictionary["scan"])
-            ]
-            clist = [
-                ("%s_crystal_%d.json" % (basepath, i), d)
-                for i, d in enumerate(dictionary["crystal"])
-            ]
-            plist = [
-                ("%s_profile_%d.json" % (basepath, i), d)
-                for i, d in enumerate(dictionary["profile"])
-            ]
-            scalelist = [
-                ("%s_scaling_model_%d.json" % (basepath, i), d)
-                for i, d in enumerate(dictionary["scaling_model"])
-            ]
-
-            # Get the list of experiments
-            edict = collections.OrderedDict(
-                [("__id__", "ExperimentList"), ("experiment", dictionary["experiment"])]
-            )
-
-            # Set paths rather than indices
-            for e in edict["experiment"]:
-                if "imageset" in e:
-                    e["imageset"] = ilist[e["imageset"]][0]
-                if "beam" in e:
-                    e["beam"] = blist[e["beam"]][0]
-                if "detector" in e:
-                    e["detector"] = dlist[e["detector"]][0]
-                if "goniometer" in e:
-                    e["goniometer"] = glist[e["goniometer"]][0]
-                if "scan" in e:
-                    e["scan"] = slist[e["scan"]][0]
-                if "crystal" in e:
-                    e["crystal"] = clist[e["crystal"]][0]
-                if "profile" in e:
-                    e["profile"] = plist[e["profile"]][0]
-                if "scaling_model" in e:
-                    e["scaling_model"] = scalelist[e["scaling_model"]][0]
-
-            to_write = (
-                ilist
-                + blist
-                + dlist
-                + glist
-                + slist
-                + clist
-                + plist
-                + scalelist
-                + [(filename, edict)]
-            )
-        else:
-            to_write = [(filename, dictionary)]
-
-        for fname, obj in to_write:
-            if compact:
-                separators = (",", ":")
-                indent = None
-            else:
-                separators = None
-                indent = 2
-            text = json.dumps(
-                obj,
-                separators=separators,
-                indent=indent,
-                ensure_ascii=True,
-                cls=AutoEncoder,
-            )
-
-            # If a filename is set then dump to file otherwise return string
-            if fname:
-                with open(fname, "w") as outfile:
-                    outfile.write(text)
-            else:
-                return text
-
-    def as_pickle(self, filename=None, **kwargs):
-        """ Dump experiment list as pickle. """
-        # Get the pickle string
-        text = pickle.dumps(self._experiment_list, protocol=pickle.HIGHEST_PROTOCOL)
-
-        # Write the file
-        if filename:
-            with open(filename, "wb") as outfile:
-                outfile.write(text)
-        else:
-            return text
-
-    def as_file(self, filename, **kwargs):
-        """ Dump experiment list as file. """
-        ext = os.path.splitext(filename)[1]
-        j_ext = [".json", ".expt"]
-        p_ext = [".p", ".pkl", ".pickle"]
-        if ext.lower() in j_ext:
-            return self.as_json(filename, **kwargs)
-        elif ext.lower() in p_ext:
-            return self.as_pickle(filename, **kwargs)
-        else:
-            ext_str = "|".join(j_ext + p_ext)
-            raise RuntimeError("expected extension {%s}, got %s" % (ext_str, ext))
+        warnings.warn(
+            "class ExperimentListDumper() is deprecated. "
+            "Use experiment_list.as_json(), experiment_list.as_pickle(), experiment_list.as_file() directly",
+            DeprecationWarning,
+            stacklevel=2,
+        )
+        assert experiment_list
+        self.as_json = experiment_list.as_json
+        self.as_pickle = experiment_list.as_pickle
+        self.as_file = experiment_list.as_file
 
 
 class ExperimentListFactory(object):
